requirements.txt
Flask
opencv-python
numpy
python-dotenv
flask-socketio
eventlet
ultralytics